Implemented -L (--local-lib-contained)!
`cpanm -L extlib Plack` will install Plack *and* its all non-core
dependencies into extlib, ready to use in your application with `use
local::lib "extlib"`

Changed how to get installed version of the modules. Instead of
running the code with backticks, it now scans the @INC and uses
MM->parse_version and version.pm for the comparison. This is so that
we can make local::lib --self-contained work correctly, and as a bonus
we can get it faster.

for perl < 5.10 there's no version->new we need to bootstrap. That
part isn't done yet. For now if you're on 5.8 you should manually
install `cpanm version` in the first run.