Dump the current @INC to a file in the build dir and include that in PERL5OPT for Makefile.PL/Build.PL so modules installed in the site/vendor lib dir won't affect Makefile.PL configuration to find prereqs.
Note that cpanminus parses Makefile.PL PREREQ anyway so as long as
Module::Install or Module::Build dumps those information this is not
strictly necessary, but when we switch completely to MYMETA and avoid
duplicating the installation check, this is a better approach.

We could dump this and switch to lib::core::only once available.