Alter tar options for non-BSD tar to ensure 'f' is last
Some older tars treat whatever immediately follows the 'f' option as the
filename, so

  tar xfz $file

is interpreted as a requst to unpack the file 'z' - instead, running

  tar zxf $file

will do the right thing.

This makes cpanm work with the tar on e.g. dd-wrt tomato (where there simply
isn't enough memory for e.g. CPAN to be usable so a working cpanm is
pretty much user's only option)