Added a "retry without parallel tests" option.
When the tests fail, if --prompt was specified, cpanm will prompt you
with a few options for how you want to proceed. This change adds an
additional option:

o) retry with test parallelism turned off

If chosen this mode manipulates the environment and then re-runs the
tests. Test suites being incompatible with parallelism is a common flaw
and this change makes it easier to proceed without resorting to using
the "look" and "force install" options.