Do not skip test phase in cpanfile for --installdeps even with --notest  Fix #222
This is a little controversial, but --notest with --installdeps means to skip
tests of the dependencies, but doesn't necessarily imply not running tests
with the *current* dist/app.

This would address the issue with Travis CI where 'test' dependencies
in META/cpanfile is ignored because of --notest option, and the actual
test will fail because test deps are not installed.

Ideally we should have an option to control which phase should be installed
via --installdeps command, but the default could be a little smart like this.