Add function to mask passwords.
I chose to use "*password*" as the mask because that is what
wget uses (or curl, I forget which).

When using wget or curl for the backend, they will emit their
own stuff to STDIN or STDOUT.  We can't easily sanitize that,
but I trust that they will not leak the password either.