Fix for git revisions that start with letter.
When git is used for a package source its sha is used for a
version. If version starts with letter it fails validation, so when
sha starts with letter this revision cannot be used in cpanfile since
it gives error like this:

Invalid metadata structure. Errors: 'af316a1' for 'version' is not a valid version. (version) [Validation: 1.0]
 at /loader/0x1178868/App/cpanminus/script.pm line 2096.

This patch forces version '9999.9999' for all git urls.