Handle META 1.x dists with dynamic_config = 0 to mean what x_dynamic_install = 1 means.
https://metacpan.org/pod/distribution/CPAN-Meta/lib/CPAN/Meta/History/Meta_1_4.pod#dynamic_config

says "[...] or whether it can be built, tested and installed solely from consulting its metadata file"
which was later removed in META spec 2.0.

This change checks if the dist was shipped only with META.yml 1.x, and treat that value as in the spec.

It is probably a risky change and should be handled as experimental with a feature flag.