fall back to skeleton if META parse fails
Some anti-social dists (e.g. AnyEvent::AIO) use JSON in their META.yml
file, which is not able to be parsed by YAML::Tiny. This results in an
empty `provides` hash, and confuses carmel.

This change falls back to the behavior of a missing META file if
CPAN::Meta is unable to parse the existing META file. Which means the
dist files are searched to populate the `provides` hash.